fix: domain, global 패키지를 sparta.jeogiyo 패키지 내부로 이동
