--- src/main/java/sparta/domain/order/controller/OrderController.java
@@ -1,4 +1,0 @@
-package sparta.domain.order.controller;
-
-public class OrderController {
-}
