--- src/main/java/sparta/domain/order/dto/request/OrderRequestDTO.java
@@ -1,4 +1,0 @@
-package sparta.domain.order.dto.request;
-
-public class OrderRequestDTO {
-}
