--- src/main/java/sparta/domain/order/dto/response/OrderResponseDTO.java
@@ -1,4 +1,0 @@
-package sparta.domain.order.dto.response;
-
-public class OrderResponseDTO {
-}
