--- src/main/java/sparta/domain/order/entity/Order.java
@@ -1,4 +1,0 @@
-package sparta.domain.order.entity;
-
-public class Order {
-}
