--- src/main/java/sparta/domain/order/repository/OrderRepository.java
@@ -1,4 +1,0 @@
-package sparta.domain.order.repository;
-
-public class OrderRepository {
-}
