--- src/main/java/sparta/domain/order/service/OrderService.java
@@ -1,4 +1,0 @@
-package sparta.domain.order.service;
-
-public class OrderService {
-}
