--- src/main/java/sparta/domain/user/UserDetailsImpl.java
@@ -1,24 +1,0 @@
-package sparta.domain.user;
-
-import java.util.Collection;
-import java.util.List;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-public class UserDetailsImpl implements UserDetails {
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of();
-    }
-
-    @Override
-    public String getPassword() {
-        return "";
-    }
-
-    @Override
-    public String getUsername() {
-        return "";
-    }
-}
