--- src/main/java/sparta/domain/user/UserDetailsServiceImpl.java
@@ -1,13 +1,0 @@
-package sparta.domain.user;
-
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return null;
-    }
-}
