--- src/main/java/sparta/domain/user/controller/UserController.java
@@ -1,8 +1,0 @@
-package sparta.domain.user.controller;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class UserController {
-
-}
