--- src/main/java/sparta/domain/user/dto/request/UserSignUpRequestDto.java
@@ -1,8 +1,0 @@
-package sparta.domain.user.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class UserSignUpRequestDto {
-
-}
