--- src/main/java/sparta/domain/user/dto/request/UserUpdateRequestDto.java
@@ -1,8 +1,0 @@
-package sparta.domain.user.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class UserUpdateRequestDto {
-
-}
