--- src/main/java/sparta/domain/user/dto/response/UserResponseDto.java
@@ -1,8 +1,0 @@
-package sparta.domain.user.dto.response;
-
-import lombok.Getter;
-
-@Getter
-public class UserResponseDto {
-
-}
