--- src/main/java/sparta/domain/user/entity/User.java
@@ -1,7 +1,0 @@
-package sparta.domain.user.entity;
-
-import sparta.global.entity.BaseTimeEntity;
-
-public class User extends BaseTimeEntity {
-
-}
