--- src/main/java/sparta/domain/user/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package sparta.domain.user.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import sparta.domain.user.entity.User;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-
-}
