--- src/main/java/sparta/domain/user/service/UserService.java
@@ -1,8 +1,0 @@
-package sparta.domain.user.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-
-}
