--- src/main/java/sparta/global/response/ApiResponse.java
@@ -1,21 +1,0 @@
-package sparta.global.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ApiResponse<T>{
-
-    private String message;
-    private T data;
-
-    public ApiResponse(String message, T data) {
-        this.message = message;
-        this.data = data;
-    }
-
-    public static <T> ApiResponse<T> of(String message, T data) {
-        return new ApiResponse<>(message, data);
-    }
-}
