--- src/main/java/sparta/global/response/CustomException.java
@@ -1,14 +1,0 @@
-package sparta.global.response;
-
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException{
-
-    private final ErrorCode errorCode;
-
-    public CustomException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-}
