--- src/main/java/sparta/global/response/GlobalExceptionHandler.java
@@ -1,39 +1,0 @@
-package sparta.global.response;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(CustomException.class)
-    protected ResponseEntity<ApiResponse<Void>> handleCustomException(CustomException customException) {
-
-        ErrorCode errorCode = customException.getErrorCode();
-
-        return ResponseEntity
-                .status(errorCode.getStatus())
-                .body(ApiResponse.of(errorCode.getMessage(), null));
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ApiResponse<Map<String, String>>> handleValidationExceptions(MethodArgumentNotValidException ex) {
-        Map<String, String> errors = new HashMap<>();
-        ex.getBindingResult().getAllErrors().forEach((error) -> {
-            String errorName = ((FieldError) error).getField();
-            String errorMessage = error.getDefaultMessage();
-            errors.put(errorName, errorMessage);
-        });
-
-        return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(ApiResponse.of("유효성 검사 실패", errors));
-    }
-}
