+++ src/main/java/sparta/jeogiyo/domain/order/controller/OrderController.java
@@ -1,0 +1,4 @@
+package sparta.jeogiyo.domain.order.controller;
+
+public class OrderController {
+}
