+++ src/main/java/sparta/jeogiyo/domain/order/dto/request/OrderRequestDTO.java
@@ -1,0 +1,4 @@
+package sparta.jeogiyo.domain.order.dto.request;
+
+public class OrderRequestDTO {
+}
