+++ src/main/java/sparta/jeogiyo/domain/order/dto/response/OrderResponseDTO.java
@@ -1,0 +1,4 @@
+package sparta.jeogiyo.domain.order.dto.response;
+
+public class OrderResponseDTO {
+}
