+++ src/main/java/sparta/jeogiyo/domain/order/entity/Order.java
@@ -1,0 +1,4 @@
+package sparta.jeogiyo.domain.order.entity;
+
+public class Order {
+}
