+++ src/main/java/sparta/jeogiyo/domain/order/repository/OrderRepository.java
@@ -1,0 +1,4 @@
+package sparta.jeogiyo.domain.order.repository;
+
+public class OrderRepository {
+}
