+++ src/main/java/sparta/jeogiyo/domain/order/service/OrderService.java
@@ -1,0 +1,4 @@
+package sparta.jeogiyo.domain.order.service;
+
+public class OrderService {
+}
