+++ src/main/java/sparta/jeogiyo/domain/user/UserDetailsServiceImpl.java
@@ -1,0 +1,13 @@
+package sparta.jeogiyo.domain.user;
+
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+
+public class UserDetailsServiceImpl implements UserDetailsService {
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return null;
+    }
+}
