+++ src/main/java/sparta/jeogiyo/domain/user/controller/UserController.java
@@ -1,0 +1,8 @@
+package sparta.jeogiyo.domain.user.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class UserController {
+
+}
