+++ src/main/java/sparta/jeogiyo/domain/user/dto/request/UserSignUpRequestDto.java
@@ -1,0 +1,8 @@
+package sparta.jeogiyo.domain.user.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class UserSignUpRequestDto {
+
+}
