+++ src/main/java/sparta/jeogiyo/domain/user/dto/response/UserResponseDto.java
@@ -1,0 +1,8 @@
+package sparta.jeogiyo.domain.user.dto.response;
+
+import lombok.Getter;
+
+@Getter
+public class UserResponseDto {
+
+}
