+++ src/main/java/sparta/jeogiyo/domain/user/entity/User.java
@@ -1,0 +1,7 @@
+package sparta.jeogiyo.domain.user.entity;
+
+import sparta.jeogiyo.global.entity.BaseTimeEntity;
+
+public class User extends BaseTimeEntity {
+
+}
