+++ src/main/java/sparta/jeogiyo/domain/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package sparta.jeogiyo.domain.user.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import sparta.jeogiyo.domain.user.entity.User;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Integer> {
+
+}
