+++ src/main/java/sparta/jeogiyo/domain/user/service/UserService.java
@@ -1,0 +1,8 @@
+package sparta.jeogiyo.domain.user.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+
+}
