+++ src/main/java/sparta/jeogiyo/global/response/ApiResponse.java
@@ -1,0 +1,21 @@
+package sparta.jeogiyo.global.response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ApiResponse<T>{
+
+    private String message;
+    private T data;
+
+    public ApiResponse(String message, T data) {
+        this.message = message;
+        this.data = data;
+    }
+
+    public static <T> ApiResponse<T> of(String message, T data) {
+        return new ApiResponse<>(message, data);
+    }
+}
