+++ src/main/java/sparta/jeogiyo/global/response/CustomException.java
@@ -1,0 +1,14 @@
+package sparta.jeogiyo.global.response;
+
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException{
+
+    private final ErrorCode errorCode;
+
+    public CustomException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+}
